refactor what-s_installed script, add options to make-world

+++ make-gwh-world.bash
@@ -3,73 +3,28 @@
 cd "$(dirname "$0")" || exit 1
 CWD=$(pwd)
 
-# a/nct6687d \
-    #     ap/aegis-rs \
-    #     ap/bat \
-    #     ap/btop \
-    #     ap/checkssl \
-    #     ap/conty-bin \
-    #     ap/delta \
-    #     ap/fd \
-    #     ap/getssl \
-    #     ap/glow \
-    #     ap/gopsuinfo \
-    #     ap/hledger-bin \
-    #     ap/jellyfin-tui \
-    #     ap/just \
-    #     ap/ledger \
-    #     ap/libtree \
-    #     ap/lsd \
-    #     ap/ncdu \
-    #     ap/nvtop \
-    #     ap/pastel \
-    #     ap/pandoc-bin \
-    #     ap/ripgrep \
-    #     ap/rkvm \
-    #     ap/starship \
-    #     ap/tinty \
-    #     ap/vivid \
-    #     ap/zauth \
-    #     d/crystal-lang \
-    #     d/factor-lang \
-    #     d/luarocks \
-    #     d/rpn-lang \
-    #     d/uxn \
-    #     d/zig-lang-bin \
-    #     e/emacs \
-    #     l/fcft \
-    #     l/libscfg \
-    #     l/notcurses \
-    #     l/spdlog \
-    #     l/tllist \
-    #     n/forgejo-bin \
-    #     n/jellyfin-bin \
-    #     xap/brightnessctl \
-    #     xap/cmd-polkit \
-    #     xap/dunst \
-    #     xap/foot \
-    #     xap/freetube \
-    #     xap/fuzzel \
-    #     xap/fuzzel-polkit-agent \
-    #     xap/gammastep \
-    #     xap/grim \
-    #     xap/guile-swayer \
-    #     xap/hpemung \
-    #     xap/i3status-rust \
-    #     xap/nextcloud-desktop \
-    #     xap/openrgb \
-    #     xap/saturnng \
-    #     xap/slurp \
-    #     xap/swappy \
-    #     xap/sway \
-    #     xap/swayidle \
-    #     xap/wbg \
-    #     xap/wlopm \
-    #     xap/wlr-randr \
-    #     xap/x48ng \
-    #     xap/x50ng \
-    #     y/csol \
-    #     y/solitaire-tui \
+DRYRUN="false"
+INSTALL="true"
+
+until [ -z "$1" ]
+do
+    case $1 in
+        "--dry-run"|"--dryrun")
+            DRYRUN="true"
+            shift
+            ;;
+        "--no-install")
+            INSTALL="false"
+            shift
+            ;;
+
+        "-h"|"--help"|*)
+            echo "$0 \\"
+            echo "  --dry-run : only check but don't build packages"
+            echo "  --no-install : don't install the packages after building"
+            exit
+    esac
+done
 
 for pkg in $("$CWD"/what-s_installed.sh | grep "^| x | " | sed 's/| x | //') ;
     do
@@ -87,10 +42,16 @@ for pkg in $("$CWD"/what-s_installed.sh | grep "^| x | " | sed 's/| x | //') ;
 
     echo -n "building /tmp/$PKGNAM > "
 
+    if [ "$DRYRUN" = "true" ]; then
+        echo "-"
+        continue
+    fi
+
     "$CWD"/make-pkg.bash "$pkg" > "$CWD/building_$PKGNAM.log" 2>&1
 
     if [ -e /tmp/"$PKGNAM" ]; then
         rm "$CWD/building_$PKGNAM.log"
+        [ "$INSTALL" = "true" ] && upgradepkg /tmp/"$PKGNAM"
         echo "✅"
     else
         echo "❌"
